Updates to circos plot for Z palustris vs O sativa
Changes were made to configuration file to add ribbons/links between Oryza sativa chromosomes (colored) and Zizania palustris chromosomes (black). The first link (test link) files are being uploaded for the first time. Positions are mostly estimates. They are based on the output of MUMmer plots. Many more links need to be added and (if possible) Z. palustris chromosome names should be altered to scaffolds because that's what they are.